add entities modificated Creation Account Controller Created repository to find songs by title, album, genre Add methods to add Song Add set Modified method and inserted an exception if the registration fields such as title and artist are not inserted Modified Song's Controller to add song. Add method to delete song to @Service Add controller to delete song Processing method to update edited method update, fixed error Add message to error in methods to update song and i try to put exception in addSong Add boolean to see if song is public or not. Add input to method addSong and try-catch-finally to close Scanner added the condition, if the song is public and the user is an administrator, he can delete the song

+++ src/main/java/com/dhmusic/DHMusic/services/SongService.java
@@ -1,5 +1,6 @@
 package com.dhmusic.DHMusic.services;
 
+import com.dhmusic.DHMusic.entities.account.entities.User;
 import com.dhmusic.DHMusic.entities.content.entities.Song;
 import com.dhmusic.DHMusic.repositories.content.repositories.SongRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -48,14 +49,18 @@ public class SongService {
         }
         }
 
-    public void deleteSong (Song song){
+    public void deleteSong (Song song) {
+        User user = new User();
         Song existSong = (Song) songRepository.findSongByTitle(song.getTitle()); //mettere indice
-        if (existSong == null){
+        if (existSong == null) {
             throw new RuntimeException("Song not exist!");
         }
-        //if() // TODO se la canzone è pubblica e l'utente è admin, cancella
-        songRepository.delete(song);
+        if (existSong.isPublic() && user.isAdmin1() == true) // TODO se la canzone è pubblica e l'utente è admin, cancella
+            songRepository.delete(song);
+        //TODO mettere eccezioni
     }
+
+
 
     public Song updateSong(Song song){
         Song existSong = (Song) songRepository.findSongById(song.getIdSong()); // id string o long?
